+++ internal/service/repository/server.go
@@ -2,8 +2,6 @@ package repository
 
 import (
 	pb "gitlab.com/gitlab-org/gitaly-proto/go"
-
-	"golang.org/x/net/context"
 )
 
 type server struct{}
@@ -11,12 +9,4 @@ type server struct{}
 // NewServer creates a new instance of a gRPC repo server
 func NewServer() pb.RepositoryServiceServer {
 	return &server{}
-}
-
-func (s *server) ApplyGitattributes(ctx context.Context, in *pb.ApplyGitattributesRequest) (*pb.ApplyGitattributesResponse, error) {
-	return nil, nil
-}
-
-func (server) FetchRemote(ctx context.Context, in *pb.FetchRemoteRequest) (*pb.FetchRemoteResponse, error) {
-	return nil, nil
 }